sid: change data types

+++ cp_benchmarks/pointers/python_sysctl.py
@@ -1,4 +1,4 @@
-from ctypes import CDLL, c_uint, byref, create_string_buffer
+from ctypes import CDLL, c_uint, c_uint64, byref
 from ctypes.util import find_library
 libc = CDLL(find_library("c"))
 
@@ -7,7 +7,7 @@ def sysctlRead(name):
     # Find out how big our buffer will be
     libc.sysctlbyname(name, None, byref(size), None, 0)
     # Make the buffer
-    value = c_uint(0)
+    value = c_uint64(0)
     # Re-run, but provide the buffer
     libc.sysctlbyname(name, byref(value), byref(size), None, 0)
     return value
@@ -18,8 +18,8 @@ def sysctlWrite(name, value):
     rv = libc.sysctlbyname(name, None, byref(size), None, 0)
     assert(rv == 0)
     # Make the buffer
-    oldValue = c_uint(0)
-    newValue = c_uint(value)
+    oldValue = c_uint64(0)
+    newValue = c_uint64(value)
     # Re-run, but provide the buffer
     rv = libc.sysctlbyname(name, 
             byref(oldValue), byref(size),
